update readme file post presentation
This site had to be rebuilt only shortly before go live of the presentation.  Original readme file was lost and so the one here before my presentation was a placeholder.  These updates should provide some insight into what this site is should anyone stumble upon it (or students want to revisit it).